requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
numpy==1.24.3
soundfile==0.12.1
librosa==0.10.1
torch==2.0.1
torchaudio==2.0.2
faster-whisper==0.9.0
livekit-server-sdk==1.0.4
pydub==0.25.1
requests==2.31.0
deepl==1.15.0
azure-cognitiveservices-speech==1.32.1
gtts==2.3.2
websockets==12.0
aiofiles==23.2.1 